read mesh data in hit shader
Currently not working, but seems close.  It renders the icosphere, but the normals don't seem correct.
I'm able to get a capture and see the buffer in nsight, and nsight (by default?) treats the buffer with padding: vec3 -> vec4.
I added std430 which should remove the padding if I understand the docs correctly.