fix reading normals in hit shader
Turns out that padding to vec4 works.  Guess that makes sense from a compute perspective -
i.e. that it's easier/faster to fetch vec4 when indexing.
So can std430 be used in the hit shader? Or perhaps I don't understand it fully.

Another bug I found is that it seems my Y axis is flipped.
Or rather, everything seems inverted.
That explains why I had to flip y when doing the camera controls,
but perhaps I have a camera bug somwhere.