fix upside down camera
Used scratchapixel site to refresh my projection matrix knowledge.

Explicitly performed conversions between different spaces in raygen.rgen.
This can be done in a simpler manner, but left as is for clarity.

Also, we can just pass in fov or tan(fov/2) info to shader to compute
correct aspect ratio scaling, and camera view vectors, instead of
passing the full inverse matrices.  Will probably do that in a later commit.