on gl_InstanceID and gl_InstanceCustomIndexNV
So, I think I now understand gl_InstanceID vs gl_InstanceCustomIndexNV.
The 2nd field of the "GeometryInstance" struct is for
gl_InstanceCustomIndexNV.  I'm currently using this in the shader
to access the instances of the TLAS.  However, I'm just incrementing
this starting from 0.  I think this should be *equivalent* to gl_InstanceID.
That is, I can just use gl_InstanceID (which should be automatically
assigned and incremented) in the shader for my use case.

What I'm still not sure about is the nonuniformEXT keyword in the
shader.  I _think_ it's asserting that the variable is not actually variable ?
Or rather that the value is known ahead of time?  Unsure ...