load vec3 normals
Aha - so figured out what I was doing wrong before.
If I specify the buffer in the shader as a vec3 buffer,
it will pad it to vec4.

Also it seems the default layout _is_ std430, and my
understanding now is that if I had vec3+float, it'll
pack those together into a vec4.

However, 16byte (vec4) *alignment* is a separate issue.

A way to get around the alignment issue is to just
have a float array and unpack explicitly.