\0/ reflections!!
surprisingly (or maybe not surprisingly?) simple, although it's
doing a mirror reflection on all non-white (or all non-red)
materials.

Will clean things up, but fun to see this easy result :)

Reflection is done as a recursive call.  This can be also be done
from the raygen shader if the hit shader returns some signal
to do the mirror reflection.

Happy New Year!!