limit trace recursion in shader
The trace recursive calls in the hit shader did not seem
to be limited by the VkRayTracingPipelineCreateInfoNV
maxRecursionDepth property.  When set to 1, it would
still do a recursive call into the trace function.

The nvpro sample limits recursion explicitly via the
ray payload data, and I've copied that as well.

This fixed the vkWaitFences device lost return value.