construct geometric normal from positions
For now just adding a separate positions buffer.
This would probably be better to have a vertex
buffer and just access the elements in the shader.

This fixes the odd shading I was seeing when
using the interpolated normal to determine
if I'm shading a backface.

?!?!  actually .... why aren't backfaces culled in
the first place?  I did not specify
VK_GEOMETRY_INSTANCE_TRIANGLE_CULL_DISABLE_BIT_NV

Normally rasterizer would cull back faces, but
it looks like we need to do this (or treat this specially)
when ray tracing.  Or perhaps I'm still missing something.