finish up refraction
This was a LOT harder than I expected.  I went through and derived
slowly the vector form of refract, as well as reviewing Snell's law
and Ray Tracing In a Weekend (RTIAW).

I knew from RTIAW that the glass ball should make the stuff
behind it appear upside down, but the objects were NOT upside
down.  After a bunch of random typing, I realized I was reversing
the Index or Refraction (IOR) ratio for outside and inside faces!?!?
That is, for outward faces, it should be outsideIOR / insideIOR,
(where outside is air, inside is glass), and inside faces the
reciprocal.  Writing my code I had it reversed.  This resulted in
the objects not being upside down, and refract hitting the
critical angle (when discriminant is less than zero) at not so
steep angles, make it look like a sphere inside a sphere.

Later I will try to model a sphere in a sphere, which should
be handled by the current code.  ... actually just tried it with
the cylinder, but it looked weird and I reverted it.  More
debugging to come.

Additionally, watching reboot red video on vulkan RTX,
I learned it's better not to do the recursion.  I'll tackle that
probably in the next commit.